Create is_prime module
