+++ pEuler.3.py
@@ -1,16 +1,5 @@
 import math
-
-
-def is_prime(n):
-    if n % 2 == 0:
-        return False
-    else:
-        t = math.sqrt(n)
-        for i in range(2, int(t)+1):
-            if n % i == 0:
-                return False
-        return True
-
+from is_prime import is_prime
 
 def prime_divisors(n):
     max_prime = 2
